+++ tests/testthat/test-target_market_share.R
@@ -647,10 +647,10 @@ test_that("w/ multiple match `level`, unweighted production is equal to ALD prod
   )
 
   ald_production <- fake_ald() %>%
-    pull(production)
+    dplyr::pull(production)
 
   out_production <- filter(out, metric == "projected") %>%
-    pull(production)
+    dplyr::pull(production)
 
   expect_equal(ald_production, out_production)
 })
@@ -694,7 +694,6 @@ test_that("with bad column errors with informative message (#267)", {
 
 test_that("`technology_share` outputs consistently when multiple
           direct_loantakers match to a single company (#265)", {
-
   matched <- fake_matched(
     id_loan = c("L1", "L2", "L3", "L4", "L5"),
     name_ald = c(rep("company a", 4), "company b")
@@ -704,7 +703,7 @@ test_that("`technology_share` outputs consistently when multiple
     mutate(name_direct_loantaker = c("company a1", "company a2", "company a3", "company a4", "company b"))
 
   ald <- fake_ald(
-    name_company = rep(c("company a","company b"), each = 2),
+    name_company = rep(c("company a", "company b"), each = 2),
     technology = rep(c("ice", "electric"), 2),
     production = c(8, 2, 15, 5)
   )
@@ -738,5 +737,57 @@ test_that("`technology_share` outputs consistently when multiple
     )
 
   expect_equal(out$technology_share, out_split_dl$technology_share)
+})
 
+test_that("`technology_share` outputs consistently when multiple
+          loans at different levels match to a single company (#265)", {
+  matched_same_level <- fake_matched(
+    id_loan = c("L1", "L2", "L3"),
+    name_direct_loantaker = c("company a1", "company a2", "company b"),
+    name_ald = c("company a", "company a", "company b")
+  )
+
+  matched_diff_level <- matched_same_level %>%
+    mutate(
+      level = c("ultimate_parent", "direct_loantaker", "ultimate_parent")
+    )
+
+  ald <- fake_ald(
+    name_company = rep(c("company a", "company b"), each = 2),
+    technology = rep(c("ice", "electric"), 2),
+    production = c(8, 2, 15, 5)
+  )
+
+  scenario <- fake_scenario(
+    technology = c("ice", "electric")
+  )
+
+  out_same_level <- target_market_share(
+    matched_same_level,
+    ald,
+    scenario,
+    region_isos_stable
+  ) %>%
+    filter(
+      metric == "projected",
+      year == 2025,
+      technology == "ice"
+    )
+
+  out_diff_level <- target_market_share(
+    matched_diff_level,
+    ald,
+    scenario,
+    region_isos_stable
+  ) %>%
+    filter(
+      metric == "projected",
+      year == 2025,
+      technology == "ice"
+    )
+
+  expect_equal(
+    out_same_level$technology_share,
+    out_diff_level$technology_share
+  )
 })