+++ NEWS.md
@@ -1,5 +1,10 @@
 # r2dii.analysis (development version)
 
+# r2dii.analysis 0.1.6
+
+* `target_market_share()` now correctly outputs `technology_share` with
+  multiple loans at different `level` to the same company (@ab-bbva #265).
+
 # r2dii.analysis 0.1.5
 
 * `target_market_share()` now errors if input `data` has an unexpected column